goon ZrStorage's WriteContextStatus

+++ drcm/zrstorage-struct.go
@@ -170,7 +170,7 @@ type Net_RoleAndContext_Data struct {
 	// 单一的绑定属性修改，1为int，2为float，3为complex
 	Single			uint8
 	// 单一的绑定修改所对应的位置，也就是0到9
-	Bit				int8
+	Bit				int
 	// 单一修改的Int
 	Int				int64
 	// 单一修改的Float
